6.1.8 Выносим общую логику обновления
В будущем нам потребуется обновлять и другие компоненты (например, статистику или фильтры) и хорошо бы переиспользовать подход, который мы уже реализовали, для этого выделим общие методы в отдельный класс Smart
